requirements/base.txt
Django==2.2.12
Pillow==7.0.0
